Revert "Added rigmor detection scream"
This reverts commit d5d7beff25fa3f62a30bba45543e6d85c5d5ffcd, reversing
changes made to 2d207855ac77ca37067b4ee09c6b7b70e2a3cc15.

+++ Assets/Scripts/Test/Viktors lekhörna/Inventory.cs
@@ -5,7 +5,7 @@ public class Inventory : MonoBehaviour
 {
     public Stack<GameObject> inventory;
     public GameObject player;
-    public float inventoryCount;
+    public int inventoryCount;
 
     private float movometer;
     private float startMovometer = 2f;
